# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublic"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
  /users/change-password:
    post:
      operationId: changePassword
      tags: [users]
      summary: Change own password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean

  # ─── Day Locks ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    idParam:
      in: path
      name: id
      required: true
      schema:
        type: string
...
  schemas:
...
    UserPublic:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, STAFF]
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
      required: [id, username, name, email, role, status]
    UserCreateInput:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, STAFF]
        password:
          type: string
      required: [username, name, email, role, password]
    UserUpdateInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, STAFF]
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
    ChangePasswordInput:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
...